add profile templates
To hide some of the complexity of generating basic resources/permissions for the short lived tokens, we are introducing a `profile-template` flag which will accept a value from the predefined list and generate the correct policies. Example to create a read only short lived token with a TTL of 15m.

```
cf-vault add my-example --profile-template "read-only" --session-duration "15m"
```

Closes #31