Detect and throw an error if the creds can't be set in the Keyring
Currently an error when setting the item misleadingly results in the success message.  There's nothing in the Keyring docs about this (`Set` doesn't return an error), but I've found that `nil` is returned when the call is successful, and an error is returned otherwise.

Unsure if this is good Golang practice or not.